Implemented login page without backend integration, only login button is set up
