+++ desk-up/src/login/login.css
@@ -1,0 +1,47 @@
+html {
+    height: 100%;
+}
+
+body {
+    align-items: center;
+    justify-content: center;
+    background-image: linear-gradient(328deg, rgba(0,0,0,1) 0%, rgba(90,60,130,1) 50%, rgba(226,206,255,1) 100%);
+    color: white;
+    font-family: Arial, Helvetica, sans-serif;
+}
+
+
+.login {
+    text-align: center;
+    padding: 200px 0;
+}
+
+
+button {
+    background-color: lightcyan;
+    border-radius: 8px;
+    border: none;
+    color: #a882b5;
+    width: 5em;
+    height: 1.5em;
+    font-family: Arial, Helvetica, sans-serif;
+    margin: 5px;
+}
+
+input {
+    border: none;
+    border-radius: 15px;
+    background: lightcyan;
+    width: 8em;
+    height: 0.25em;
+    font-family: Arial, Helvetica, sans-serif;
+    padding: 20px;
+    margin: 10px;
+}
+
+button:hover {
+    border: 1px solid paleturquoise;
+    background-color: white;
+    color: rgb(99, 52, 110);
+  }
+
